tests/requirements-test.txt
# Testing dependencies for TeleDrive
# Install with: pip install -r tests/requirements-test.txt

# Core testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-html>=3.1.0

# Flask testing
pytest-flask>=1.2.0

# Database testing
pytest-postgresql>=4.1.0
pytest-sqlite>=1.0.0

# Performance testing
pytest-benchmark>=4.0.0
locust>=2.14.0

# Security testing
bandit>=1.7.0
safety>=2.3.0

# Code quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0

# Coverage reporting
coverage>=7.0.0
codecov>=2.1.0

# Mock and testing utilities
responses>=0.23.0
factory-boy>=3.2.0
faker>=18.0.0
freezegun>=1.2.0

# Web testing
selenium>=4.8.0
playwright>=1.30.0

# API testing
requests>=2.28.0
httpx>=0.24.0

# Image processing for test data
Pillow>=9.4.0

# Additional utilities
parameterized>=0.9.0
testfixtures>=7.1.0
